Add Unsquezee operation
As in other operations involving indices, we add both - Julia-friendly and ONNX-friendly versions. ONNX-friendly version follows the semantics in ONNX, e.g. 0-based row-major indexing, applies conversions and calls the Julia-friendly version. It's much easier to encapsulate conversions into the op itself rather than generate a separate operations on the tape/graph for it, but in future we will need to support both somehow