Add Slice
This implementation covers basic scenarios for Slice, which I expect to make 80-90% of cases. It does not yet handle negative start and end indices, nor implements corretness checks for input. This is intentional - even though tests pass, I'm still not sure the conversion between ONNX and Julia is correct, thus I prefer to keep things simple until it's proved to be correct at least in basic cases in a large real-world model